Move get_message_timestamp common helper method
